+++ kernel/tr/strings/e_string.cpp
@@ -165,7 +165,7 @@ void estr::copy_text_estr(xptr dest, xptr src, str_off_t count)
     int dest_spc_blk = E_STR_BLK_FREE_SPACE(dest_blk);
 
     CHECKP(src);
-    int src_spc_blk = BLK_BEGIN_INT(XADDR(src)) + PAGE_SIZE - (int)XADDR(src);
+    int src_spc_blk = BLK_BEGIN_INT(XADDR(src)) + PAGE_SIZE - (uintptr_t)XADDR(src);
     int real_count = s_min(dest_spc_blk, s_min(src_spc_blk, count));
 
     memcpy(executor_globals::e_string_buf, XADDR(src), real_count);
@@ -269,7 +269,7 @@ int estr_cursor::copy_blk(char *buf)
     if (!m_count) return 0;
 
     CHECKP(m_str);
-    int len = BLK_BEGIN_INT(XADDR(m_str)) + PAGE_SIZE - (int)XADDR(m_str);
+    int len = BLK_BEGIN_INT(XADDR(m_str)) + PAGE_SIZE - (uintptr_t)XADDR(m_str);
     if (len >= m_count)
     {
         len = m_count;
@@ -290,7 +290,7 @@ int estr_cursor::get_blk(char **ptr)
     if (!m_count) return 0;
 
     CHECKP(m_str);
-    int len = BLK_BEGIN_INT(XADDR(m_str)) + PAGE_SIZE - (int)XADDR(m_str);
+    int len = BLK_BEGIN_INT(XADDR(m_str)) + PAGE_SIZE - (uintptr_t)XADDR(m_str);
     if (len >= m_count)
     {
         len = m_count;
@@ -313,7 +313,7 @@ void estr_feed(string_consumer_fn fn, void *p, xptr src, str_off_t count) // or 
 	{
 		CHECKP(src);
 
-		int src_spc_blk = BLK_BEGIN_INT(XADDR(src)) + PAGE_SIZE - (int)(XADDR(src));
+		int src_spc_blk = BLK_BEGIN_INT(XADDR(src)) + PAGE_SIZE - (uintptr_t)(XADDR(src));
 		int real_count = s_min(src_spc_blk, count);
 
 		fn((char*)XADDR(src), real_count, p);
@@ -332,7 +332,7 @@ void estr_copy_to_buffer(char *dest, xptr src, str_off_t count)
 	{
 		CHECKP(src);
 
-		int src_spc_blk = BLK_BEGIN_INT(XADDR(src)) + PAGE_SIZE - (int)(XADDR(src));
+		int src_spc_blk = BLK_BEGIN_INT(XADDR(src)) + PAGE_SIZE - (uintptr_t)(XADDR(src));
 		int real_count = s_min(src_spc_blk, count);
 
 		memcpy(dest, XADDR(src), real_count);
